// src/app/core/api/generated/schemas/netParcelRateDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */
import type { NetParcelRateDtoMaxDeliveryDate } from "./netParcelRateDtoMaxDeliveryDate";
import type { NetParcelRateDtoMinDeliveryDate } from "./netParcelRateDtoMinDeliveryDate";

export interface NetParcelRateDto {
  /** @nullable */
  min_delivery_date?: NetParcelRateDtoMinDeliveryDate;
  /** @nullable */
  max_delivery_date?: NetParcelRateDtoMaxDeliveryDate;
  service_name: string;
  service_code: string;
  total_price: string;
  currency: string;
}